requirements.txt
fastapi
uvicorn[standard]
playwright
fastkml
shapely
psycopg2-binary
sqlalchemy
geoalchemy2
pydantic
